requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
idna==3.3
importlib-metadata==4.10.0
Jinja2==3.0.3
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
wasabi==0.9.0
zipp==3.6.0
